+++ package.json
@@ -18,8 +18,8 @@
   },
   "dependencies": {
     "apollo": "^0.2.2",
-    "apollo-cache-inmemory": "^1.1.9",
-    "apollo-client": "^2.2.1",
+    "apollo-cache-inmemory": "^1.1.10",
+    "apollo-client": "^2.2.6",
     "apollo-client-preset": "^1.0.6",
     "apollo-link-http": "^1.3.2",
     "bootstrap": "4.0.0-beta.2",
@@ -28,7 +28,6 @@
     "dropzone": "^5.2.0",
     "graphql": "^0.13.1",
     "graphql-tag": "^2.6.1",
-    "jest-enzyme": "^4.2.0",
     "jquery": "^3.1.1",
     "jquery.cookie": "^1.4.1",
     "jss": "^9.6.0",
@@ -68,16 +67,16 @@
     "@babel/preset-es2015": "^7.0.0-beta.40",
     "@babel/preset-react": "^7.0.0-beta.40",
     "@babel/preset-stage-0": "^7.0.0-beta.40",
+    "@types/jest": "^22.2.0",
     "@types/react-dom": "^16.0.4",
     "@types/react-router-dom": "^4.2.4",
+    "@types/react-test-renderer": "^16.0.1",
     "autoprefixer": "^7.1.4",
     "babel-eslint": "^8.2.2",
     "babel-jest": "^22.2.2",
     "babel-loader": "^8.0.0-beta.2",
     "chai": "^4.1.2",
     "css-loader": "^0.28.10",
-    "enzyme": "^3.3.0",
-    "enzyme-adapter-react-16": "^1.1.1",
     "eslint": "^4.18.1",
     "eslint-config-prettier": "^2.9.0",
     "eslint-config-standard": "^10.2.1",
@@ -94,9 +93,10 @@
     "node-sass": "^4.7.2",
     "postcss": "^6.0.11",
     "postcss-loader": "^2.0.6",
-    "react-test-renderer": "^15.6.1",
+    "react-test-renderer": "^16.2.0",
     "regenerator-runtime": "^0.11.1",
     "sass-loader": "^6.0.6",
+    "ts-jest": "^22.4.1",
     "ts-loader": "^4.0.0",
     "webpack": "^4.0.1",
     "webpack-bundle-tracker": "^0.2.1",
@@ -121,13 +121,25 @@
     }
   },
   "jest": {
-    "globals": {}
+    "globals": {},
+    "transform": {
+      "^.+\\.tsx?$": "ts-jest"
+    },
+    "testRegex": "(/__tests__/.*|(\\.|/)(test|spec))\\.(jsx?|tsx?)$",
+    "moduleFileExtensions": [
+      "ts",
+      "tsx",
+      "js",
+      "jsx",
+      "json",
+      "node"
+    ]
   },
   "scripts": {
     "build-assets": "node ./node_modules/webpack/bin/webpack.js -p --progress",
     "heroku-postbuild": "npm run build-assets && npm run build-emails",
     "start": "node ./node_modules/webpack/bin/webpack.js -d --watch --progress",
     "build-emails": "mkdir -p templates/templated_email/compiled/ && find templates -name \"*.mjml\" -not -path \"templates/templated_email/source/*partials/*\" -not -path \"templates/templated_email/source/*shared/*\" -exec ./node_modules/.bin/mjml -l skip {} -o templates/templated_email/compiled/ \\;",
-    "test": "jest saleor/static/dashboard/js/components/app/test"
+    "test": "jest saleor/static/dashboard-next/js/"
   }
 }
